:pencil: rename: 기존 post 액션 함수 삭제 및 컴포넌트 경로 재설정
